requirements-dev.txt
freezegun
pytest
pytest-django
pytest-pythonpath
sure
tox
